backend/extractor/requirements.txt
fastapi
uvicorn
python-multipart
google-genai
pillow
python-dotenv
pydantic